[wip] rearrange files to allow importing runner independently
