+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "ncc build src/index.js -o dist",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "local": "node test/local.js"
   },
   "keywords": [],
   "author": "",
