+++ src/action.js
@@ -1,0 +1,12 @@
+import { getInput } from '@actions/core';
+import FileSet from 'file-set';
+
+export async function run() {
+  const glob = getInput('FILES');
+  const fileSet = new FileSet();
+  await fileSet.add([glob]);
+  const { files } = fileSet; 
+  
+  const gist = getInput('GIST');
+  console.log(`uploading ${files} to ${gist}`);
+}
